// frontend/src/components/FormInputValues/FormInputValues.types.ts
export interface FormInputsTypes {
    initialWireSize: number;
    finalWireSize: number;
    totalTransitions: number;
    firstReduction: number;
    drawingVelocity: number;
    initialBlockNumber: number;
    lastDieReduction: number;
    lastBlockNumber: number;
    steelGrade: string;
    patentedTensileStrengthMin: number;
    patentedTensileStrengthMax: number;
}